requirements.txt
rawpy>=0.19.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
flask>=3.0.0
exifread>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
opencv-python>=4.8.0
face-recognition>=1.3.0
